# openapi.yaml
openapi: 3.0.1
servers:
  - url: 'http://m0.internal.diabetes.watch'
info:
  description: >-
    Manage a cluster of processes varied by environment.  One process will
    be run against multiple environments, one process per environment.
    You can learn more abou [multienv](https://github.com/bewest/node-multienv)
  version: 0.0.1
  title: MultiEnv
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: bewest@gmail.com
  license:
    name: Affero GPL 3.0
tags:
  - name: cluster
    description: Manages processes by cluster.
  - name: proc
    description: Access to proc managed by cluster.
  - name: env
    description: Environment for a process.
paths:
  /cluster:
    get:
      tags:
        - cluster
      summary: Get list of managed clusters
      description: JSON description of all currently running processes.
      operationId: get_cluster_list
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proc'
        '400':
          description: Invalid status value
  '/cluster/{id}':
    get:
      tags:
        - proc
      summary: Running proc by id.
      description: Returns details of a process by `id`.
      operationId: get_cluster_by_id
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proc'
        '400':
          description: Invalid ID supplied
        '404':
          description: Process not found
      parameters:
        - in: path
          name: id
          description: "id of cluster."
          schema:
            type: string
  /history:
    get:
      tags:
        - proc
      summary: Get managed processes.
      description: >-
        Returns description of all processes seen by `multienv`.  Useful to get
        info on processes that may have died and/or been deleted.
      operationId: get_history
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proc'

      parameters: []
  /environs:
    get:
      tags:
        - env
      summary: Get list of managed environments.
      description: JSON description of all managed environments.
      operationId: get_environments_list
      parameters: []
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proc'
  '/environs/{name}':
    parameters:
      - in: path
        name: name
        description: "name of managed environment."
        schema:
          type: string
    post:
      tags:
        - env
      summary: Create new environment
      description: >-
        **POST** a json object to define a new environment.

        `/environs/:name` returns a json description of environment by `name`.

        This is the name of the environment file relative to `WORKER_ENV`, sans
        `.env`

        extension.

        Returns new environment variables as a json object.
      operationId: create_environment
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proc'
        '400':
          description: Invalid Order
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Proc/properties/custom_env'
        description: Environment to run defined process.
        required: true
      parameters: []
    get:
      summary: 'Get Named Environment'
      description: >-
        /environs/:name returns a json description of environment by name. This
        is the name of the environment file relative to WORKER_ENV, sans .env
        extension.
      parameters: []
      operationId: get_environment_by_name
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proc'
        default:
          description: Default response
    delete:
      summary: Remove an environment by name.
      description: >-
        `/environs/:name` - Delete (and stop a running process for) an
        environment.

        This is the name of the environment file relative to `WORKER_ENV`, sans
        `.env`

        extension.
      parameters: []
      operationId: delete_environment_by_name
      responses:
        default:
          description: Default response
  '/environs/{name}/env':
    get:
      tags:
        - proc
      summary: Get environment variables as a json object.
      description: >-
        `/environs/:name/env`

        This is the `name` of the environment file relative to `WORKER_ENV`,
        sans

        `.env` extension.
      operationId: get_custom_env_by_name
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proc/properties/custom_env'
      parameters:
        - in: path
          name: name
          description: "Name of environment"
          schema:
            type: string
  '/environs/{name}/env/{field}':
    get:
      tags:
        - env
      summary: Get value of environment variable of `field`.
      description: >-
        `/environs/:name/env/:field` - retrieve value of a single environment
        variable.

        This is the `name` of the environment file relative to `WORKER_ENV`,
        sans

        `.env` extension.

        The `field` is the name of the single environment variable.
      operationId: get_environment_variable
      parameters:
        - in: path
          name: name
          description: "Name of environment."
          schema:
            type: string
        - in: path
          name: field 
          description: "Name of environment variable."
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proc'
        '400':
          description: Invalid ID supplied
        '404':
          description: Order not found
externalDocs:
  description: See AsyncAPI example
  url: 'https://mermade.github.io/shins/asyncapi.html'
components:
  schemas:

    Proc:
      description: Describe a worker process.
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: id of worker.
        state:
          type: string
          description: state of worker.
        is_dead:
          name: "isDead"
          type: boolean
          description: whether or not process is dead.
        custom_env:
          type: object
          description: Map of environment variables applied to worker.


  links: {}
  callbacks: {}
security: []